# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Referral API",
        "description": "API for managing referrals between business units",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        },
        {
            "url": "http://mytotalhealth.com.my/referral-api",
            "description": "MyHealth Server (Staging)"
        }
    ],
    "paths": {
...
        "/api/external-referees": {
            "get": {
                "tags": [
                    "External Referees"
                ],
                "summary": "Get list of external referees",
                "operationId": "326e3e1f0772f6069319d357f8974cae",
                "responses": {
                    "200": {
                        "description": "List of external referees",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ExternalReferee"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "External Referees"
                ],
                "summary": "Create a new external referee",
                "operationId": "88d1a51076805019aaff7a64f02b613b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "position": {
                                        "type": "string"
                                    },
                                    "specialty": {
                                        "type": "string"
                                    },
                                    "external_organization_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "organization": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "address": {
                                                "type": "string"
                                            },
                                            "postcode": {
                                                "type": "string"
                                            },
                                            "state": {
                                                "type": "string"
                                            },
                                            "country": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "External referee created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExternalReferee"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/external-referees/{referee}": {
            "get": {
                "tags": [
                    "External Referees"
                ],
                "summary": "Get external referee details",
                "operationId": "ae88010ab5465e711f40bc3e9880bb8c",
                "parameters": [
                    {
                        "name": "referee",
                        "in": "path",
                        "description": "External referee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "External referee details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExternalReferee"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "External referee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "External Referees"
                ],
                "summary": "Update an external referee",
                "operationId": "95bb654ad87646c4bd2c9e454c910aa5",
                "parameters": [
                    {
                        "name": "referee",
                        "in": "path",
                        "description": "External referee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "phone": {
                                        "type": "string"
                                    },
                                    "position": {
                                        "type": "string"
                                    },
                                    "specialty": {
                                        "type": "string"
                                    },
                                    "external_organization_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "organization": {
                                        "properties": {
                                            "name": {
                                                "type": "string"
                                            },
                                            "address": {
                                                "type": "string"
                                            },
                                            "postcode": {
                                                "type": "string"
                                            },
                                            "state": {
                                                "type": "string"
                                            },
                                            "country": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "External referee updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExternalReferee"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Validation failed"
                                        },
                                        "errors": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "External referee not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "External Referees"
                ],
                "summary": "Delete an external referee",
                "operationId": "f651ea635355730989eebc94100687e0",
                "parameters": [
                    {
                        "name": "referee",
                        "in": "path",
                        "description": "External referee ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "External referee deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "External referee deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "External referee not found"
                    },
                    "500": {
                        "description": "Server error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ExternalReferee": {
                "required": [
                    "name",
                    "email",
                    "phone",
                    "position",
                    "specialty"
                ],
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "external_organization_id": {
                        "type": "integer",
                        "nullable": true
                    },
                    "name": {
                        "type": "string"
                    },
                    "email": {
                        "type": "string"
                    },
                    "phone": {
                        "type": "string"
                    },
                    "position": {
                        "type": "string"
                    },
                    "specialty": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "boolean",
                        "default": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "deleted_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "organization": {
                        "oneOf": [
                            {
                                "$ref": "#/components/schemas/ExternalOrganization"
                            }
                        ],
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "FormDetail": {
                "required": [
                    "form_detail_id",
                    "field_name",
                    "field_type",
                    "is_required"
                ],
                "properties": {
                    "form_detail_id": {
                        "type": "integer",
                        "example": 101
                    },
                    "field_name": {
                        "type": "string",
                        "example": "Blood Pressure"
                    },
                    "field_type": {
                        "type": "string",
                        "example": "text"
                    },
                    "is_required": {
                        "type": "boolean",
                        "example": true
                    },
                    "field_value": {
                        "oneOf": [
                            {
                                "type": "string",
                                "example": "120/80"
                            },
                            {
                                "type": "array",
                                "items": {
                                    "properties": {
                                        "form_detail_id": {
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "field_value": {
                                            "type": "string",
                                            "example": "Option A"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        ]
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
            "ExternalOrganization": {
                "required": [
                    "name",
                    "address",
                    "postcode",
                    "state",
                    "country"
                ],
                "type": "object",
                "allOf": [
                    {
                        "$ref": "#/components/schemas/FormDetail"
                    },
                    {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "Sample Organization"
                            },
                            "address": {
                                "type": "string",
                                "example": "123 Main Street"
                            },
                            "postcode": {
                                "type": "string",
                                "example": "12345"
                            },
                            "state": {
                                "type": "string",
                                "example": "California"
                            },
                            "country": {
                                "type": "string",
                                "example": "United States"
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time"
                            },
                            "deleted_at": {
                                "type": "string",
                                "format": "date-time",
                                "nullable": true
                            }
                        },
                        "type": "object"
                    }
                ]
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter JWT Bearer token in the format: Bearer {token}"
            }
        }
    },
...
}